restructured the folder layout for tests
