+++ TerraformCLI/src/tests/terraform-apply/apply-azurerm-with-invalid-auth-scheme.ts
@@ -1,12 +1,13 @@
-import { TaskScenario } from './scenarios';
-import { TerraformInputs } from './scenarios-terraform';
-import './scenarios-terraform'
+import { TaskScenario } from '../scenarios';
+import { TerraformInputs } from '../scenarios-terraform';
+import '../scenarios-terraform';
 
 const environmentServiceName = "dev";
 const subscriptionId: string = "sub1";
 const tenantId: string = "ten1";
 const servicePrincipalId: string = "servicePrincipal1";
 const servicePrincipalKey: string = "servicePrincipalKey123";
+const authScheme: string = 'foo';
 
 const expectedEnv: { [key: string]: string } = {
     'ARM_SUBSCRIPTION_ID': subscriptionId,
@@ -16,13 +17,12 @@ const expectedEnv: { [key: string]: string } = {
 }
 
 const terraformCommand: string = "apply";
-const commandArgs: string = '-auto-approve';
 
 new TaskScenario<TerraformInputs>()
-    .withAzureRmServiceEndpoint(environmentServiceName, subscriptionId, tenantId, servicePrincipalId, servicePrincipalKey)
+    .withAzureRmServiceEndpoint(environmentServiceName, subscriptionId, tenantId, servicePrincipalId, servicePrincipalKey, authScheme)
     .inputTerraformCommand(terraformCommand)
     .withInputs({ environmentServiceName: environmentServiceName })
     .answerTerraformExists()
-    .answerTerraformCommandIsSuccessful(commandArgs)
+    .answerTerraformCommandIsSuccessful()
     .answerTerraformCommandWithVarsFileAsWorkingDirFails()
     .run()
