+++ TerraformCLI/src/tests/terraform-apply/apply-azurerm-with-var-file.ts
@@ -1,13 +1,12 @@
-import { TaskScenario } from './scenarios';
-import { TerraformInputs } from './scenarios-terraform';
-import './scenarios-terraform'
+import { TaskScenario } from '../scenarios';
+import { TerraformInputs } from '../scenarios-terraform';
+import '../scenarios-terraform'
 
 const environmentServiceName = "dev";
 const subscriptionId: string = "sub1";
 const tenantId: string = "ten1";
 const servicePrincipalId: string = "servicePrincipal1";
 const servicePrincipalKey: string = "servicePrincipalKey123";
-const authScheme: string = 'foo';
 
 const expectedEnv: { [key: string]: string } = {
     'ARM_SUBSCRIPTION_ID': subscriptionId,
@@ -16,15 +15,15 @@ const expectedEnv: { [key: string]: string } = {
     'ARM_CLIENT_SECRET': servicePrincipalKey,
 }
 
-const terraformCommand: string = "plan";
+const terraformCommand: string = "apply";
+const commandArgs: string = '-auto-approve';
 
 new TaskScenario<TerraformInputs>()
-    .withAzureRmServiceEndpoint(environmentServiceName, subscriptionId, tenantId, servicePrincipalId, servicePrincipalKey, authScheme)
+    .withAzureRmServiceEndpoint(environmentServiceName, subscriptionId, tenantId, servicePrincipalId, servicePrincipalKey)
     .inputTerraformCommand(terraformCommand)
     .withInputs({ environmentServiceName: environmentServiceName })
+    .inputTerraformVarsFile('foo.vars')
     .answerTerraformExists()
-    .answerTerraformCommandIsSuccessful()
+    .answerTerraformCommandIsSuccessful(commandArgs)
     .answerTerraformCommandWithVarsFileAsWorkingDirFails()
     .run()
-
-
