+++ TerraformCLI/src/tests/terraform-init/init-with-backend-invalid.ts
@@ -1,6 +1,6 @@
-import { TaskScenario } from './scenarios';
-import { TerraformInputs } from './scenarios-terraform';
-import './scenarios-terraform'
+import { TaskScenario } from '../scenarios';
+import { TerraformInputs } from '../scenarios-terraform';
+import '../scenarios-terraform'
 
 new TaskScenario<TerraformInputs>()
     .inputTerraformCommand('init')
