--- TerraformCLI/src/tests/terraform-not-exists.ts
@@ -1,8 +1,0 @@
-import { TaskScenario } from './scenarios';
-import { TerraformInputs } from './scenarios-terraform';
-import './scenarios-terraform'
-
-new TaskScenario<TerraformInputs>()
-    .inputTerraformCommand("version")
-    .answerTerraformExists(false)
-    .run();
