+++ TerraformCLI/src/tests/terraform-validate/validate-with-var-file.ts
@@ -1,9 +1,10 @@
-import { TaskScenario } from './scenarios';
-import { TerraformInputs } from './scenarios-terraform';
-import './scenarios-terraform'
+import { TaskScenario } from '../scenarios';
+import { TerraformInputs } from '../scenarios-terraform';
+import '../scenarios-terraform'
 
 export let validateWithNoArgs = new TaskScenario<TerraformInputs>()
     .inputTerraformCommand("validate")
+    .inputTerraformVarsFile("foo.vars")
     .answerTerraformExists()
     .answerTerraformCommandIsSuccessful()
     .answerTerraformCommandWithVarsFileAsWorkingDirFails()
