+++ TerraformCLI/src/tests/terraform/terraform-not-exists.ts
@@ -1,0 +1,8 @@
+import { TaskScenario } from '../scenarios';
+import { TerraformInputs } from '../scenarios-terraform';
+import '../scenarios-terraform'
+
+new TaskScenario<TerraformInputs>()
+    .inputTerraformCommand("version")
+    .answerTerraformExists(false)
+    .run();
